+++ CHANGELOG.md
@@ -6,6 +6,7 @@ on [Keep a CHANGELOG](http://keepachangelog.com/). This project adheres to
 ## Unreleased
 ### Added
 - GeneratorFactory interface to allow factory creation of Generators
+- GeneratorProperties interface for custom implementation of property class
 
 ## [1.7.0] - 2017-08-17
 ### Added